clean up pyserial base dsdf
